+++ src/service/records-service.js
@@ -1,4 +1,4 @@
-import { RecordsRepository } from '../repository/records.repository.js';
+import { RecordsRepository } from '../repository/records-repository.js';
 import axios from 'axios';
 
 export class RecordsService {
